Add a C startup stub
The program will run from a C piece of code where main() is
located. The C++ main() was renamed to "RunMainCode" and a
"MainCPP" to be called by this new C main is added, which
swallows all C++ exceptions thrown by the C++ code.

I need this because it turns out that threads can only be
properly handled from C code. C++ is not thread compatible
because its "operator new()" is not thread safe or at least
it is not on my setup (Ubuntu Linux, g++ 4.8.4). The idea
here is to have only one thread that runs the C++ code and
all the other threads doing only support stuff like
handling packets.