Remove "runtime debug stacks" kludge
This kludge relies on putting special directives in the form
"DSTACK(some-text-to-be-shown-in-stackframe)" at the
beginning of every procedure and function. Things that
require me to "do manual thing X for each element in the
source code" I hate with passion so this thing has to go.
Additionally, I noticed that this "debug stack dumping
facility" already started to rot - most procedures and
functions lack these special directives.