Convert configuration file reading to C
I am going to need access to the configuration file data from
the C code and a C++ class does not cut it for this. Thus I
need the configuration data to be read in C.

This code performs limited parsing of boolean values and
integers. Everything else is stored in the configuration file
data structure as strings.